chore: remove duplicacy
